Add files via upload
Here is a MATLAB script that allows the user to hard code a function toward finding the center of mass of the corresponding plane under the function's curve. 
This script uses methods from calculus 2. That is, it does not use calculus 3 methods, namely this code does not find the double integral (or triple for a 3D object).